+++ packages/blocks-editor/src/index.ts
@@ -1,4 +1,2 @@
 export * from './Editor/BlocksEditor';
 export * from './Editor/BlocksEditorComposer';
-export * from './Editor/Commands';
-export * from './Editor/ClassNames';