+++ packages/icons/src/Icons/index.ts
@@ -46,8 +46,8 @@ import ColorFillIcon from './ic-color-fill.svg'
 import CopyIcon from './ic-copy.svg'
 import CreateAccountIllustration from './create-account-illustration.svg'
 import DashboardIcon from './ic-dashboard.svg'
-import DiamondIcon from './diamond-with-horizontal-lines.svg'
 import DiamondFilledIcon from './ic-diamond-filled.svg'
+import DiamondIcon from './diamond-with-horizontal-lines.svg'
 import DownloadIcon from './ic-download.svg'
 import DragIcon from './ic-drag.svg'
 import DrawIcon from './ic-draw.svg'
@@ -167,12 +167,14 @@ import SNLogoFull from './ic-sn-logo-full.svg'
 import SNLogoIcon from './ic-standard-notes.svg'
 import SortDescendingIcon from './ic-sort-descending.svg'
 import SpreadsheetsIcon from './ic-spreadsheets.svg'
-import StarFilledIcon from './ic-star-filled.svg'
 import StarCircleFilled from './ic-star-circle-filled.svg'
+import StarFilledIcon from './ic-star-filled.svg'
 import StarIcon from './ic-star.svg'
 import StarVariantFilledIcon from './ic-star-variant-filled.svg'
 import StrikethroughIcon from './ic-strikethrough.svg'
+import SubscriptIcon from './ic-subscript.svg'
 import SubtractIcon from './ic-subtract.svg'
+import SuperscriptIcon from './ic-superscript.svg'
 import SyncIcon from './ic-sync.svg'
 import TasksIcon from './ic-tasks.svg'
 import TextCircleIcon from './ic-text-circle.svg'
@@ -373,7 +375,9 @@ export {
   StarIcon,
   StarVariantFilledIcon,
   StrikethroughIcon,
+  SubscriptIcon,
   SubtractIcon,
+  SuperscriptIcon,
   SyncIcon,
   TasksIcon,
   TextCircleIcon,