+++ packages/icons/src/Lexical/index.ts
@@ -1,0 +1,61 @@
+import LexicalCaretRightFill from './caret-right-fill.svg'
+import LexicalCheck from './square-check.svg'
+import LexicalCode from './code.svg'
+import LexicalHorizontalRule from './horizontal-rule.svg'
+import LexicalLink from './link.svg'
+import LexicalListOL from './list-ol.svg'
+import LexicalListUL from './list-ul.svg'
+import LexicalQuote from './chat-square-quote.svg'
+import LexicalTable from './table.svg'
+import LexicalTextCenter from './text-center.svg'
+import LexicalTextJustify from './justify.svg'
+import LexicalTextLeft from './text-left.svg'
+import LexicalTextParagraph from './text-paragraph.svg'
+import LexicalTextRight from './text-right.svg'
+import LexicalTweet from './tweet.svg'
+import LexicalYoutube from './youtube.svg'
+import TypeBold from './type-bold.svg'
+import TypeH1 from './type-h1.svg'
+import TypeH2 from './type-h2.svg'
+import TypeH3 from './type-h3.svg'
+import TypeH4 from './type-h4.svg'
+import TypeH5 from './type-h5.svg'
+import TypeH6 from './type-h6.svg'
+import TypeItalic from './type-italic.svg'
+import TypeStrikethrough from './type-strikethrough.svg'
+import TypeSubscript from './type-subscript.svg'
+import TypeSuperscript from './type-superscript.svg'
+import TypeUnderline from './type-underline.svg'
+import LexicalPencilFill from './pencil-fill.svg'
+
+export {
+  LexicalCaretRightFill,
+  LexicalCheck,
+  LexicalCode,
+  LexicalHorizontalRule,
+  LexicalLink,
+  LexicalListOL,
+  LexicalListUL,
+  LexicalQuote,
+  LexicalTable,
+  LexicalTextCenter,
+  LexicalTextJustify,
+  LexicalTextLeft,
+  LexicalTextParagraph,
+  LexicalTextRight,
+  LexicalTweet,
+  LexicalYoutube,
+  TypeBold,
+  TypeH1,
+  TypeH2,
+  TypeH3,
+  TypeH4,
+  TypeH5,
+  TypeH6,
+  TypeItalic,
+  TypeStrikethrough,
+  TypeSubscript,
+  TypeSuperscript,
+  TypeUnderline,
+  LexicalPencilFill,
+}