+++ packages/web/src/javascripts/Components/NoteView/NoteView.tsx
@@ -8,7 +8,6 @@ import ProtectedItemOverlay from '@/Components/ProtectedItemOverlay/ProtectedIte
 import { ElementIds } from '@/Constants/ElementIDs'
 import { PrefDefaults } from '@/Constants/PrefDefaults'
 import { StringDeleteNote, STRING_DELETE_LOCKED_ATTEMPT, STRING_DELETE_PLACEHOLDER_ATTEMPT } from '@/Constants/Strings'
-import { featureTrunkEnabled, FeatureTrunkName } from '@/FeatureTrunk'
 import { log, LoggingDomain } from '@/Logging'
 import { debounce, isDesktopApplication, isMobileScreen } from '@/Utils'
 import { classNames } from '@/Utils/ConcatenateClassNames'
@@ -996,7 +995,7 @@ class NoteView extends AbstractComponent<NoteViewProps, State> {
     const renderHeaderOptions = isMobileScreen() ? !this.state.plaintextEditorFocused : true
 
     const editorMode =
-      featureTrunkEnabled(FeatureTrunkName.Blocks) && this.note.noteType === NoteType.Blocks
+      this.note.noteType === NoteType.Blocks
         ? 'blocks'
         : this.state.editorStateDidLoad && !this.state.editorComponentViewer && !this.state.textareaUnloading
         ? 'plain'
@@ -1091,7 +1090,9 @@ class NoteView extends AbstractComponent<NoteViewProps, State> {
                 </div>
               )}
             </div>
-            <LinkedItemBubblesContainer linkingController={this.viewControllerManager.linkingController} />
+            {editorMode !== 'blocks' && (
+              <LinkedItemBubblesContainer linkingController={this.viewControllerManager.linkingController} />
+            )}
           </div>
         )}
 
@@ -1115,7 +1116,7 @@ class NoteView extends AbstractComponent<NoteViewProps, State> {
           ) : null}
 
           {editorMode === 'component' && this.state.editorComponentViewer && (
-            <div className="component-view">
+            <div className="component-view flex-grow">
               <ComponentView
                 key={this.state.editorComponentViewer.identifier}
                 componentViewer={this.state.editorComponentViewer}
@@ -1148,7 +1149,13 @@ class NoteView extends AbstractComponent<NoteViewProps, State> {
 
           {editorMode === 'blocks' && (
             <div className={classNames('blocks-editor w-full flex-grow overflow-hidden overflow-y-scroll')}>
-              <BlockEditor key={this.note.uuid} application={this.application} note={this.note} />
+              <BlockEditor
+                key={this.note.uuid}
+                application={this.application}
+                note={this.note}
+                linkingController={this.viewControllerManager.linkingController}
+                filesController={this.viewControllerManager.filesController}
+              />
             </div>
           )}