+++ packages/desktop/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.23.290](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.289...@standardnotes/desktop@3.23.290) (2022-11-10)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.23.289](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.288...@standardnotes/desktop@3.23.289) (2022-11-09)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.23.288](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.287...@standardnotes/desktop@3.23.288) (2022-11-09)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.23.287](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.286...@standardnotes/desktop@3.23.287) (2022-11-09)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.23.286](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.285...@standardnotes/desktop@3.23.286) (2022-11-09)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.23.285](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.284...@standardnotes/desktop@3.23.285) (2022-11-09)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.23.284](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.283...@standardnotes/desktop@3.23.284) (2022-11-09)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
 ## [3.23.283](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.282...@standardnotes/desktop@3.23.283) (2022-11-08)
 
 **Note:** Version bump only for package @standardnotes/desktop