+++ packages/files/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.12.8](https://github.com/standardnotes/app/compare/@standardnotes/files@1.12.7...@standardnotes/files@1.12.8) (2022-11-09)
+
+**Note:** Version bump only for package @standardnotes/files
+
 ## [1.12.7](https://github.com/standardnotes/app/compare/@standardnotes/files@1.12.6...@standardnotes/files@1.12.7) (2022-11-07)
 
 **Note:** Version bump only for package @standardnotes/files